# openapi.yaml
openapi: 3.1.0
info:
  title: Cadence API
  version: "1.0.0"
  summary: REST API for the Cadence training platform.
  description: >
    A REST API over the same data model that powers the Cadence app — activities
    and their 1 Hz streams, designed workouts, scheduling, athlete thresholds and
    zones, coaching, and gear. JSON over HTTPS, OAuth 2.0, predictable resource URLs.


    ## Authentication

    OAuth 2.0 authorization code flow. Send the access token as a bearer header on
    every request:


    `Authorization: Bearer cad_at_9f2a7c…`


    ## Pagination

    List endpoints are cursor-paginated. Pass `cursor` from the previous response's
    `next_cursor`, plus an optional `limit` (max 200).


    ## Query language (CQL)

    The `q` parameter on `GET /v1/activities` accepts the same query language as the
    Activities search bar — field comparisons (`>` `<` `>=` `<=` `=` `!=`), boolean
    `AND` / `OR`, and `ORDER BY`. Example:
    `sport = run AND tag ~ race AND distance > 42.2km`.
  contact:
    name: Cadence Developer Platform
    url: https://developers.cadence.cc
  license:
    name: Proprietary
servers:
  - url: https://api.cadence.cc
    description: Production
security:
  - oauth2: []
  - bearerAuth: []
...
paths:
...
  /v1/athletes/{id}:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Athletes & zones]
      summary: Retrieve an athlete
      operationId: getAthlete
      responses:
        "200":
          description: Profile and thresholds.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Athlete"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Athletes & zones]
      summary: Update thresholds
      operationId: updateAthlete
      description: >
        Updates profile and threshold values. Changing a threshold recomputes every
        zone set that references it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                age: { type: integer }
                weight_kg: { type: number, description: Used for the Avg Watts/kg stat on Activity Analysis. }
                ftp:
                  type: integer
                  description: >
                    Functional threshold power (W). A manually-entered value is
                    recorded as a new threshold_history entry unconditionally (no
                    outlier check - an explicit human declaration, not a computed
                    candidate) and ages out after threshold_window_days like any
                    other entry. A no-op if it matches the latest ledger entry
                    already, since this form resubmits every field on every save.
                critical_run_power: { type: integer, description: Critical running power (W). Same threshold_history semantics as ftp. }
                threshold_pace: { type: string, description: "mm:ss per km. Same threshold_history semantics as ftp." }
                threshold_window_days: { type: integer, description: See the User schema. }
                threshold_sanity_pct: { type: integer, description: See the User schema. }
                threshold_warning_days: { type: integer, enum: [14, 21, 28, 0], description: See the User schema. }
                lthr: { type: integer, description: Lactate threshold HR (bpm). }
                max_hr: { type: integer }
                resting_hr:
                  type: integer
                  description: >
                    Optional; only feeds the Karvonen HRR% stat, not a zone
                    reference threshold, so changing it never appears in
                    zones_recomputed.
      responses:
        "200":
          description: Updated athlete.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Athlete"
                  - type: object
                    properties:
                      zones_recomputed:
                        type: array
                        items: { type: string }
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://api.cadence.cc/oauth/authorize
          tokenUrl: https://api.cadence.cc/oauth/token
          scopes:
            activities:read: Read activities and streams
            activities:write: Upload and edit activities
            workouts:write: Create and edit workouts
            calendar:write: Schedule workouts
            coach: Access an athlete roster
            gear:write: Manage gear
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AthleteId:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
...
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            type: { type: string }
            code: { type: string }
            message: { type: string }
            param: { type: [string, "null"] }
...
    User:
      type: object
      properties:
        id: { type: string, example: usr_neil }
        name: { type: string, example: Neil Goodgame }
        email: { type: string, example: neil@example.cc }
        email_verified:
          type: boolean
          description: >
            True immediately for social signups (the provider already verified the
            address). False for a password signup until they redeem the link sent by
            POST /v1/auth/register, or resend one via POST /v1/auth/resend-verification.
          example: true
        age: { type: integer, example: 34 }
        weight_kg: { type: number, example: 72 }
        ftp: { type: integer, example: 265 }
        critical_run_power: { type: integer, example: 285 }
        running_power_source:
          type: string
          enum: [stryd, native]
          default: stryd
          description: >-
            Which running-power reading to trust when a FIT file carries both a watch's own
            native estimate (e.g. Garmin Running Power) and a Stryd footpod reading - native
            algorithms tend to read meaningfully higher than Stryd for the same effort, so the
            non-selected source is completely ignored, not used as a fallback.
          example: stryd
        threshold_pace: { type: string, example: "4:00" }
        threshold_window_days:
          type: integer
          default: 112
          description: >
            Trailing window (days) a threshold's best qualifying effort must fall
            within to count - see GET /v1/athletes/{id}/thresholds. Default 112 (16
            weeks).
          example: 112
        threshold_sanity_pct:
          type: integer
          default: 30
          description: >
            Per-candidate outlier filter - an effort implying more than this %
            deviation from the then-current value is excluded (e.g. corrupt
            power-meter data).
          example: 30
        threshold_warning_days:
          type: integer
          enum: [14, 21, 28, 0]
          default: 21
          description: >
            How many days before a threshold's current source activity ages out of
            threshold_window_days the Threshold suggestions feature (GET
            /v1/athletes/{id}/threshold-suggestions) warns the athlete. 0 = off.
            Scaled down for threshold_window_days < 84 (a quarter of the window is
            used instead of this raw setting).
          example: 21
        lthr: { type: integer, example: 155 }
        max_hr: { type: integer, example: 186 }
        resting_hr:
          type: [integer, "null"]
          description: Optional. Only used for the Karvonen heart-rate-reserve % shown on Activity Analysis's Stats tab.
          example: 48
        is_coach:
          type: boolean
          description: Whether this user also coaches other users. A user can train and coach at once.
          example: true
    Athlete:
      allOf:
        - $ref: "#/components/schemas/User"
      description: Deprecated alias of User, kept for compatibility.
...